moved Host default implementation into ContextTr
